requirements.txt
awscli==1.16.92
botocore==1.12.82
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-prometheus==1.0.15
djangorestframework==3.9.1
docutils==0.14
dumb-init==1.2.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
prometheus-client==0.5.0
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
